Moved Nitro Runtime manager to it's own crate (more platforms pending)
