+++ nitro-runtime-manager/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "nitro_runtime_manager"
+authors = ["The Veracruz Development Team"]
+description = "The trusted Veracruz runtime's command-and-control module for AWS Nitro Enclaves. Co-ordinates and directs the rest of the runtime"
+edition = "2018"
+version = "0.1.0"
+
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+nix = { version = "0.20.2" }
+nsm_api = { git = "https://github.com/aws/aws-nitro-enclaves-nsm-api.git/", tag = "v0.2.0", package = "aws-nitro-enclaves-nsm-api"}
+nsm_lib = { git = "https://github.com/aws/aws-nitro-enclaves-nsm-api.git/", tag = "v0.2.0", package = "nsm-lib"}
+raw-fd = { git = "https://github.com/veracruz-project/nitro-enclave", tag = "v0.1" }
+runtime-manager = { path = "../runtime-manager", features = ["nitro"] }
+rustix = "=0.37.13"
+veracruz-utils = { path = "../veracruz-utils", features = ["linux"] }
