+++ runtime-manager/src/lib.rs
@@ -13,5 +13,7 @@
 #![cfg_attr(any(feature = "icecap"), feature(rustc_private))]
 #![cfg_attr(any(feature = "icecap"), feature(format_args_nl))]
 
+pub mod common_runtime;
 pub mod managers;
 pub mod platforms;
+pub mod platform_runtime;
