+++ runtime-manager/src/platforms/mod.rs
@@ -13,5 +13,3 @@
 pub mod icecap;
 #[cfg(feature = "linux")]
 pub mod linux;
-#[cfg(feature = "nitro")]
-pub mod nitro;
