+++ workspaces/nitro-runtime/Makefile
@@ -9,7 +9,7 @@
 # See the `LICENSE_MIT.markdown` file in the Veracruz root director for licensing
 # and copyright information.
 
-.PHONY: all clean clean-cargo-lock clippy default doc fmt nitro runtime-manager-enclave
+.PHONY: all clean clean-cargo-lock clippy default doc fmt nitro nitro-runtime-manager
 
 default: all
 
@@ -22,23 +22,23 @@ include $(WORKSPACE_DIR)/shared.mk
 
 all: nitro
 
-nitro: runtime-manager-enclave
+nitro: nitro-runtime-manager
 	$(MAKE) runtime_manager.eif PCR0
 
-runtime_manager.eif PCR0 &: target/$(ARCH)-unknown-linux-musl/$(PROFILE_PATH)/runtime-manager-enclave crates/runtime-manager/dockerdir/Dockerfile
+runtime_manager.eif PCR0 &: target/$(ARCH)-unknown-linux-musl/$(PROFILE_PATH)/nitro_runtime_manager crates/nitro-runtime-manager/dockerdir/Dockerfile
 	rm -rf docker
 	mkdir -p docker
-	cp target/$(ARCH)-unknown-linux-musl/$(PROFILE_PATH)/runtime-manager-enclave docker
-	cp crates/runtime-manager/dockerdir/Dockerfile docker
+	cp target/$(ARCH)-unknown-linux-musl/$(PROFILE_PATH)/nitro_runtime_manager docker
+	cp crates/nitro-runtime-manager/dockerdir/Dockerfile docker
 	perl -i -pe "s/x86_64/$(ARCH)/" docker/Dockerfile
-	nitro-cli build-enclave --docker-dir docker --docker-uri runtime_manager --output-file runtime_manager.eif > measurements.json
+	nitro-cli build-enclave --docker-dir docker --docker-uri runtime_manager --output-file nitro_runtime_manager.eif > measurements.json
 	cat measurements.json | jq -r '.Measurements.PCR0' > PCR0
 	rm -rf docker
 
-runtime-manager-enclave:
+nitro-runtime-manager:
 	rustup target add $(ARCH)-unknown-linux-musl
 	cargo build --target $(ARCH)-unknown-linux-musl $(PROFILE_FLAG) $(V_FLAG)
-	strip -p target/$(ARCH)-unknown-linux-musl/$(PROFILE_PATH)/runtime-manager-enclave
+	strip -p target/$(ARCH)-unknown-linux-musl/$(PROFILE_PATH)/nitro_runtime_manager
 
 doc:
 	cargo doc
@@ -46,7 +46,7 @@ doc:
 clippy:
 	rustup target add $(ARCH)-unknown-linux-musl
 	cargo clippy --target $(ARCH)-unknown-linux-musl $(PROFILE_FLAG) $(V_FLAG) \
-		-p nitro-runtime -p runtime-manager -p execution-engine \
+		-p nitro-runtime -p nitro_runtime_manager -p execution-engine \
 		-p session-manager -p policy-utils -p platform-services \
 		-- --no-deps
